refactor(web): update path
