+++ packages/web/src/app/(authenticated)/inbox/components/message/item/app-type.ts
