+++ packages/web/src/app/(authenticated)/inbox/components/message/item/index.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image';
 import { buttonVariants } from '~/components/ui/button';
 import { cn } from '~/libs/classes';
 import type { Message } from '~/libs/types/message';
-import { GenerateMessageDialog } from '../gen-message/dialog';
+import { GenerateMessageDialog } from '../../gen-message/dialog';
 import AccountIcon from './account-icon';
 import { type AppProps, AppsProps, defaultAppProps } from './app-type';
 import Label from './label';
