+++ packages/web/src/app/(authenticated)/inbox/components/message/item/label-type.ts
